added minimal viable test
